backend_files/requirements.txt
fastapi
uvicorn[standard]
jpush
pydantic
python-dotenv
fastmcp
